Move coefficient tests to the reader
